moved remaining config class definitions to config.py
